web-dev-projects/CongressDB/requirements.txt
certifi==2024.2.2
charset-normalizer==2.0.12
idna==3.6
python-dotenv==1.0.1
requests==2.26.0
urllib3==1.26.18
